Bumping up number of read threads to 4
I've been going back and forth on this. The original goal of cutting
down the number of read threads was to slow the rate of queue growth.
But the queue limit is a better way of achieving this goal. Having
a larger limit therefore makes it easier for the script to breeze
through the "fast track" list.